Move listening functionality to a separate shared component
This commit will create the Listenable component. The Tracker will
inherit its functionality.
